backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
sqlparse==0.5.3
Pillow==10.1.0
pytesseract==0.3.10
pandas==2.1.4
openpyxl==3.1.2
python-dateutil==2.8.2
django-cors-headers==4.3.1
